Implement special enumerators in Row and Table
When someone tries to enumerate a table directly (not Table.rows), they
probably mean to enumerate rows. This adds a simple wrapper to do that.

When enumerating a Row, enumerate both the columns and fields Collections
simultaneously, where the first param (key) is the column name, and the
second param (value) is the field's value: `for col_name, value in Row`