Fix performance issue in Collections.Count()
This originally looped through the entire object to get the number of
items. For large objects, this could be a huge performance hit. Instead,
use `this.SetCapacity(0)`, which returns the object's capacity after
shrinking the object to fit. Any performance loss due to shrinking the
object prematurely is gained back exponentially by removing a loop of ahk
code.

Consider:

    while this.Count() < 1000000
            this.insert(A_Index)

This code would loop 500,000,500,000 extra times before this commit.
Insane.